add my-council-account page (#9)
This is the page where people can learn about what to expect in registration or they could log in if they already had an account and if the log in button actually worked (since there is no log in function yet the log in button doesn't work). If they click on Register then they will move to the register-requirements page.